quest app implementation version 12: Corrected ExceptionHandlerServlet: all html code were substituted by exception.jsp Added styles_exception.css used by exception.jsp Added v1 of Readme Applied autoformatting
