+++ README.md
@@ -1,9 +1,98 @@
 # javarush-quest
 
+## Краткое описание
+
+Данный проект - это текстовый квест. Вопрос на каждом следующем шаге зависит от предыдущего ответа.
+
+При запуске приложения пользователю открывается приветственная страница.
+На этой странице пользователю предлагается ввести своё имя, чтобы начать игру.
+
+>Приветственная страница:
+>
+>![img.png](img.png)
+
+После того, как игрок ввел своё имя и нажал на кнопку [Introduce yourself], запускается квест.
+
+>Первый вопрос квеста:
+>
+>![img_1.png](img_1.png)
+
+В случае, если игрок не указал свое имя, но нажал на кнопку [Introduce yourself], квест предупредит пользователя о том, что имя не может быть пустым. 
+
+>Обработка ошибки:
+>
+>![img_7.png](img_7.png)
+
+Задача игрока выбрать один из представленных вариантов и нажать на кнопку [Answer].
+
+Если игрок не выбрал ни один из ответов, приложение сгенерирует алерт для юзера, подсказывающее ему сделать выбор.
+
+>Алерт:
+>
+>![img_6.png](img_6.png)
+
+В зависимоти от выбранного ответа, игра приводит к выигрышу или проигрышу.
+
+При любом из реультатов игрок может начать квест заново, а также посмотреть статистику.
+
+>Окно статистики:
+>
+>![img_2.png](img_2.png)
+
+# Запуск проекта из Docker:
+
 Deploy war file on Tomcat Docker Container
-- Install [Docker](https://docs.docker.com/install).
-- Clone this repository $git clone https://github.com/akursekova/javarush-quest.git
-- cd javarush-quest # from your root directory
-- $docker build -t javarush-quest .
-- $docker run -p 7777:8080 javarush-quest
+- Установить [Docker](https://docs.docker.com/install)
+- Перейти по ссылке https://github.com/akursekova/javarush-quest/tree/dev and нажать Donwnload ZIP
+- Имя скачанного проекта будет javarush-quest-dev, так как проект скачан из dev ветки
+- Открыть Terminal и запустить команду cd javarush-quest-dev # из root directory
+- $docker build -t javarush-quest-dev .
+- $docker run -p 7777:8080 javarush-quest-dev
 - http://localhost:7777/javarush-quest/quest
+
+>Приложение запущенное на http://localhost:7777/javarush-quest/quest:
+>
+>![img_4.png](img_4.png)
+
+# Краткое описание классов
+
+В корневом пакете проекта `dev.akursekova.quest` хранятся классы:
+- AppServletContextListener
+- ExceptionHandlerServlet
+- QuestServlet,  
+
+а также пакеты:
+- repository
+- service
+- subjects
+
+В пакете `subjects` содержатся классы, описывающие основные объекты данного квеста:
+- User
+- Question
+- Answer
+
+- `User`: описывает общие поля и методы, принадлежащие игроку, принимающему участие в игре.
+- `Question`, `Answer`: отвечает за поля и методы вопросов/ответов, появляющихся в игре.
+
+В пакете `repository` хранится класс `UserRepository`, ответственный за хранение всех пользователей в системе.
+
+В пакете `service` хранится `QuestionService` - класс, в котором хранится всё дерево вопросов квеста.
+Метод `init` всегда возвращает первый вопрос квеста. 
+
+Класс `AppServletContextListener` - это listener, который во время запуска приложения создает все экземпляры сервисов и помещает из в сервлет контекст.
+
+Класс `QuestServlet` - основной класс данного квеста. В методах `doGet` и `doPost` реализована основная бизнес-логика:
+- логика отображения вопросов/ответов
+- перемещение между jsp страницами
+- проверка выбрал ли игр какой-либо ответ
+
+Класс `ExceptionHandlerServlet`, ответственный за обработку ошибок. 
+
+В папке `WEB-INF` хранятся все jsp страницы проекта:
+- `index.jsp`: jsp, описывающее приветственную страницу
+- `question.jsp`: jsp, в которой описано отображение вопросов, ответов, в также кнопки [Answer], [Restart], [Statistics]
+- `exception.jsp`: jsp, на которую происходит редирект из ExceptionHandlerServlet в момент, когда в квесте произошла какая-либо ошибка
+
+В папке `css` хранятся стили, используемые jsp страницами
+В папке `img` хранится картинка, используемая для фона квеста
+
