+++ src/main/java/dev/akursekova/quest/ExceptionHandlerServlet.java
@@ -1,0 +1,23 @@
+package dev.akursekova.quest;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@WebServlet(name = "ExceptionHandlerServlet", value = "/exception_handler")
+public class ExceptionHandlerServlet extends HttpServlet {
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        processError(request, response);
+    }
+
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        processError(request, response);
+    }
+
+    private void processError(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
+        request.getRequestDispatcher("/WEB-INF/exception.jsp").forward(request, response);
+    }
+}
