+++ src/main/java/dev/akursekova/quest/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package dev.akursekova.app.repository;
+package dev.akursekova.quest.repository;
 
-import dev.akursekova.app.subjects.User;
+import dev.akursekova.quest.subjects.User;
 
 import java.util.HashMap;
 
@@ -8,13 +8,13 @@ public class UserRepository {
     HashMap<String, User> users = new HashMap<>();
 
     public void add(User user) {
-        if (user.getName().equals("")){
+        if (user.getName().equals("")) {
             throw new IllegalArgumentException("user name cannot be empty");
         }
         users.put(user.getName(), user);
     }
 
-    public boolean exists(String userName){
+    public boolean exists(String userName) {
         return users.containsKey(userName);
     }
