+++ src/main/java/dev/akursekova/quest/subjects/Question.java
@@ -1,4 +1,4 @@
-package dev.akursekova.app.subjects;
+package dev.akursekova.quest.subjects;
 
 import lombok.Builder;
 import lombok.Data;
