+++ src/main/resources/log4j2.xml
@@ -10,14 +10,14 @@
                 <Pattern>%d %p %c{1.} [%t] %m%n</Pattern>
             </PatternLayout>
             <Policies>
-                <TimeBasedTriggeringPolicy />
+                <TimeBasedTriggeringPolicy/>
                 <SizeBasedTriggeringPolicy size="10 MB"/>
             </Policies>
         </RollingFile>
 
     </Appenders>
     <Loggers>
-        <Logger name="dev.akursekova.app.QuestServlet" level="trace" additivity="false">
+        <Logger name="dev.akursekova.quest.QuestServlet" level="trace" additivity="false">
             <AppenderRef ref="LogToRollingFile"/>
         </Logger>
         <Root level="error">
