+++ src/main/webapp/WEB-INF/question.jsp
@@ -1,5 +1,5 @@
-<%@ page import="dev.akursekova.app.subjects.Question" %>
-<%@ page import="dev.akursekova.app.subjects.User" %>
+<%@ page import="dev.akursekova.quest.subjects.Question" %>
+<%@ page import="dev.akursekova.quest.subjects.User" %>
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
 <%@ taglib prefix="fmt" uri="http://java.sun.com/jsp/jstl/fmt" %>
 <%@ taglib prefix="fn" uri="http://java.sun.com/jsp/jstl/functions" %>
@@ -10,7 +10,6 @@
 <head>
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link rel="stylesheet" type="text/css" href="<c:url value='/css/styles.css' />">
-    <!-- Import bootstrap cdn -->
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css">
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.0/jquery.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
@@ -49,30 +48,21 @@
                 </c:otherwise>
             </c:choose>
 
-           <%--<%
-                String user = (String) session.getAttribute("user").getName();
-                String ipAddress = (String) session.getAttribute("ipAddress");
-                Integer numberOfGames = (Integer) session.getAttribute("numberOfGames");
-            %>--%>
-
             <%
                 User user = (User) session.getAttribute("user");
                 Integer numberOfGames = user.getNumberOfGames();
                 String ipAddress = user.getIpAddress();
             %>
 
-            <!-- Trigger the modal with a button -->
             <button type="button" class="btn btn-primary btn-md" id="myBtn">Statistics</button>
 
-            <!-- Modal -->
+
             <div class="modal fade" id="myModal" role="dialog" data-backdrop="false">
                 <div class="modal-dialog">
-                    <!-- Modal content-->
                     <div class="modal-content">
                         <div class="modal-header">
                             Statistics
                             <button type="button" class="close" data-dismiss="modal">×</button>
-                            <%--                <h6 class="modal-title">Statistics</h6>--%>
                         </div>
                         <div class="modal-body">
                             <p>
@@ -81,12 +71,9 @@
                                 IP address: <b><%=ipAddress %>
                             </b> <br>
                                 Completed games: <b><%=numberOfGames %>
-                           </b> <br>
+                            </b> <br>
                             </p>
                         </div>
-<%--                        <div class="modal-footer">--%>
-<%--                            <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>--%>
-<%--                        </div>--%>
                     </div>
                 </div>
             </div>
