+++ src/main/webapp/WEB-INF/web.xml
@@ -4,10 +4,10 @@
 
 <web-app>
     <display-name>Archetype Created Web Application</display-name>
-<!--    <context-param>-->
-<!--        <param-name>log4jConfiguration</param-name>-->
-<!--        <param-value>/Users/alinakursekova/IdeaProjects/javarush-quest/src/main/resources/log4j2.xml</param-value>-->
-<!--    </context-param>-->
+    <!--    <context-param>-->
+    <!--        <param-name>log4jConfiguration</param-name>-->
+    <!--        <param-value>/Users/alinakursekova/IdeaProjects/javarush-quest/src/main/resources/log4j2.xml</param-value>-->
+    <!--    </context-param>-->
 
     <error-page>
         <exception-type>java.lang.Throwable</exception-type>
